+++ src/main/data/CR7.txt
@@ -2,3 +2,7 @@
 [D][ ] siuuu (by: Jan 1 2001 06:00 PM)
 [D][ ] run (by: Dec 5 2001 06:00 PM)
 [D][ ] jump (by: Dec 5 2001 06:00 PM)
+[T][ ] run
+[T][ ] run
+[T][ ] run
+[T][ ] runn
