+++ src/main/java/commands/Command.java
@@ -1,3 +1,7 @@
+package commands;
+
+import functions.*;
+
 import java.io.IOException;
 
 public abstract class Command {
