+++ src/main/java/commands/ErrorCommand.java
@@ -1,3 +1,7 @@
+package commands;
+
+import functions.*;
+
 public class ErrorCommand extends Command {
     protected String errorMsg;
     public ErrorCommand(String errorMsg) {
