+++ src/main/java/commands/ListCommand.java
@@ -1,3 +1,7 @@
+package commands;
+
+import functions.*;
+
 import java.io.IOException;
 
 public class ListCommand extends Command {
