+++ src/main/java/commands/MarkCommand.java
@@ -1,14 +1,19 @@
+package commands;
+
+import functions.*;
+import tasks.*;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     protected int num;
-    public UnmarkCommand(int num) {
+    public MarkCommand(int num) {
         this.num = num;
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
-        Task t = tasks.unmarkTask(this.num);
-        ui.showUnmarkMsg(t);
+        Task t = tasks.markTask(this.num);
+        ui.showMarkMsg(t);
         storage.saveFiles(tasks.showList());
     }
 }
