+++ src/main/java/commands/UnmarkCommand.java
@@ -1,14 +1,19 @@
+package commands;
+
+import functions.*;
+import tasks.*;
+
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
     protected int num;
-    public DeleteCommand(int num) {
+    public UnmarkCommand(int num) {
         this.num = num;
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
-        Task t = tasks.deleteTask(num);
-        ui.showDeleteMsg(t, tasks.numOfTasks());
+        Task t = tasks.unmarkTask(this.num);
+        ui.showUnmarkMsg(t);
         storage.saveFiles(tasks.showList());
     }
 }
