+++ src/main/java/functions/Storage.java
@@ -1,3 +1,7 @@
+package functions;
+
+import tasks.Task;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
